Moved payment_utils module into new app, payments. Updated code base to reflect this.
