+++ wildlifelicensing/apps/applications/templates/wl/process/process_app.html
@@ -103,7 +103,7 @@
                                     <p class="no-margin">
                                         {{ data.payment_status }}
                                         {% if data.payment_status == 'Awaiting Payment' %}
-                                            <a href="{% url 'wl_main:manual_payment' data.application.id %}?redirect_url={{ request.build_absolute_uri }}"
+                                            <a href="{% url 'wl_payments:manual_payment' data.application.id %}?redirect_url={{ request.build_absolute_uri }}"
                                                 class="pull-right">Enter Payment</a>
                                         {% endif %}
                                     </p>
