+++ wildlifelicensing/apps/dashboard/views/base.py
@@ -7,7 +7,7 @@ import logging
 from dateutil.parser import parse as date_parse
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.staticfiles.templatetags.staticfiles import static
-from django.core.urlresolvers import reverse_lazy, reverse
+from django.core.urlresolvers import reverse
 from django.db.models import Q
 from django.db.models.query import EmptyQuerySet
 from django.shortcuts import redirect
@@ -19,7 +19,9 @@ from ledger.licence.models import LicenceType
 from wildlifelicensing.apps.applications.models import Application
 from wildlifelicensing.apps.dashboard.forms import LoginForm
 from wildlifelicensing.apps.main.helpers import is_officer, is_assessor, render_user_name
-from wildlifelicensing.apps.main.payment_utils import get_application_payment_status, PAYMENT_STATUS_AWAITING, \
+
+
+from wildlifelicensing.apps.payments.utils import get_application_payment_status, PAYMENT_STATUS_AWAITING, \
     PAYMENT_STATUSES
 
 logger = logging.getLogger(__name__)
@@ -74,7 +76,7 @@ def render_payment(application, redirect_url):
     result = '{}'.format(PAYMENT_STATUSES[status])
     if status == PAYMENT_STATUS_AWAITING:
         url = '{}?redirect_url={}'.format(
-            reverse('wl_main:manual_payment', args=[application.id]),
+            reverse('wl_payments:manual_payment', args=[application.id]),
             redirect_url
         )
         result += ' <a href="{}">Enter payment</a>'.format(url)
