+++ wildlifelicensing/apps/main/urls.py
@@ -6,7 +6,6 @@ from wildlifelicensing.apps.main.views import ListProfilesView, CreateProfilesVi
     DeleteProfileView, IdentificationView, EditAccountView, SearchCustomersView, ListDocumentView, \
     EditDocumentView, CreateDocumentView ,DeleteDocumentView, LicenceRenewalPDFView, \
     CommunicationsLogListView, AddCommunicationsLogEntryView, BulkLicenceRenewalPDFView
-from wildlifelicensing.apps.main.payment_utils import CheckoutApplicationView, ManualPaymentView, PaymentsReportView
 
 
 urlpatterns = [
@@ -37,9 +36,5 @@ urlpatterns = [
 
     # general communications log
     url('^add-log-entry/([0-9]+)/$', AddCommunicationsLogEntryView.as_view(), name='add_log_entry'),
-    url('^log-list/([0-9]+)/$', CommunicationsLogListView.as_view(), name='log_list'),
-
-    url('^checkout_application/([0-9]+)/$', CheckoutApplicationView.as_view(), name='checkout_application'),
-    url('^manual_payment/([0-9]+)/$', ManualPaymentView.as_view(), name='manual_payment'),
-    url('^payments_report/?$', PaymentsReportView.as_view(), name='payments_report'),
+    url('^log-list/([0-9]+)/$', CommunicationsLogListView.as_view(), name='log_list')
 ]
