+++ wildlifelicensing/apps/payments/apps.py
@@ -1,0 +1,9 @@
+from __future__ import unicode_literals
+
+from django.apps import AppConfig
+
+
+class PaymentsConfig(AppConfig):
+    name = 'wildlifelicensing.apps.payments'
+    label = 'wl_payments'
+    verbose_name = 'WL Payments'
