+++ wildlifelicensing/apps/reports/templates/wl/reports.html
@@ -111,7 +111,7 @@
         <div class="row">
             <div class="col-md-12">
                 <h3>Payments</h3>
-                <form method="get" id="payments-form" action="{% url 'wl_main:payments_report' %}">
+                <form method="get" id="payments-form" action="{% url 'wl_payments:payments_report' %}">
                     {% csrf_token %}
                     <div class="row">
                         <div class="col-md-4">
