+++ wildlifelicensing/urls.py
@@ -14,6 +14,7 @@ urlpatterns = [
     url(r'^customer_management/', include('wildlifelicensing.apps.customer_management.urls', namespace='wl_customer_management')),
     url(r'^reports/', include('wildlifelicensing.apps.reports.urls', namespace='wl_reports')),
     url(r'^returns/', include('wildlifelicensing.apps.returns.urls', namespace='wl_returns')),
+    url(r'^payments/', include('wildlifelicensing.apps.payments.urls', namespace='wl_payments')),
 ] + ledger_patterns
 
 if settings.DEBUG:
